requirements_test.txt
requests
mysql-connector-python
flask
uwsgidecorators
redis
pyyaml
pytest
jsonpickle